Django-Admin/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.1.0
bootstrap-admin==0.4.4
click==8.0.4
Deprecated==1.2.13
Django==4.0.2
django-mdeditor==0.1.20
django-redis==5.2.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dotenv==0.19.2
redis==4.1.4
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
wrapt==1.13.3
